Refactor project name from "gbot" to "gitscribe"
